G: Modification de substract pour ajouter le support des nombres à virgules

+++ src/stringcalculator.py
@@ -15,7 +15,7 @@ class StringCalculator :
         sub = None
         for part in parts :
             try :
-                part = int(part)
+                part = float(part)
                 if part > 1000 :
                     part = 0
             except ValueError :
